# PLY-62: Music Library Performance Optimization (4/6 requirements complete)